Gave up trying to make not-variant functionality more consistent.
I was trying to make e.g. .whereJsonNotEquals() to return all rows which didn't match .whereJsonEquals() method call with same parameters, but it did appear to be more complicated as I expected. Maybe that feature could be added in some future version if someone finds good way to implement it.