Update Insert type definition
Since it accepts an object but returns an instance of the Model. This is helpful in cases where we strictly define the properties of the instance but, allow creating the instance without some of this values. e.g.:

```ts
const accountInstance = await Account.query().insert({
    name: "Pepe"
});
await accountInstance.instanceMethod();
```

This example will throw in typescript since the instanceMethod is not part of the passed object, but after the patch, it works.