Update ModelOrObject to Partial<T>
As far as I can tell, the functions that accept ModelOrObject parameters
(Insert<T>, InsertGraphAndFetch<T>, insertAndFetch<T>. update*, patch,
relate) all allow Object as well as Model so that the caller can provide
an object that does not define all the properties of the Model.

This seems like a weaker type check than is possible with the Partial<T>
type that is provided by typescript:

https://www.typescriptlang.org/docs/handbook/advanced-types.html#mapped-types

This commit changes the ModelOrObject and ModelsOrObjects types to
PartialModel and PartialOrPartials types and makes appropriate
replacements.