Add tabbed examples for different ES versions
I had to use a combination of vue components and a plugin for the
vuepress default theme. Vuepress's default config has the power to be
configured heavily, but it's a little weird. If I just used the vue
components in the md directly, it would take the children of the
components (which should be turned into codeblocks) and make that just
one text component, so it wouldn't get processed by markdown-it. But if
I used markdown-it-container, the HTML replacement occurs before
runtime.

It's possible using <ClientOnly> would have worked, but I didn't try.
This way feels more extensible, though slightly gross.