Fix insert graph method typing
Previously, the SingleQueryBuilder definition went first, so even if an array was passed, it would be typed as SingleQueryBuilder. This created all sorts of weird effects, like making it impossible to write a type predicate on a model subclass.

This commit reorders it so it tries to type it as ArrayQueryBuilder first. It also adds a test to make sure it is typed correctly.